Rename FCL source files per style guide
