+++ drake/systems/plants/collision/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(fcl)
 
 if (fcl_FOUND)
   #add_definitions( -DBULLET_COLLISION -DBT_USE_DOUBLE_PRECISION )
-  set( drakeCollision_SRC_FILES ${drakeCollision_SRC_FILES} FCLModel.cpp )
+  set( drakeCollision_SRC_FILES ${drakeCollision_SRC_FILES} fcl_model.cc )
   set(fcl "yes" CACHE STRING "yes" )
 endif()
