+++ drake/systems/plants/collision/fcl_model.h
